use the same gauss filter for deconv methods
In deconvf the gauss filter is defined as exp(-0.5*(f/f1)**2),
the default is f1=0.5 corresponding to a cut-off frequency in Hz.
In deconv_iter it is defined as exp(-0.25*(2*pi*f/f0)**2) without the dt term,
the default width is f0=3.

The relation between f1 and f0 is f1=f0/(2**0.5 * pi),
so the default in deconv_iter of f0=3 corresponds to f1=0.67
(being the cut-off frequency of the low-pass filter).

I now switched to the same definition as in deconvf and also use the same default value 0.5.